+++ Pivot Table/Compact Layout/NET Standard/Compact Layout/Compact Layout.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34310.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Compact Layout", "Compact Layout\Compact Layout.csproj", "{93AEF7B0-2D67-4814-B451-351538E248DE}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{93AEF7B0-2D67-4814-B451-351538E248DE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{93AEF7B0-2D67-4814-B451-351538E248DE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{93AEF7B0-2D67-4814-B451-351538E248DE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{93AEF7B0-2D67-4814-B451-351538E248DE}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F614D7FE-9DB7-489B-B4D0-6B6B3E323209}
+	EndGlobalSection
+EndGlobal
