+++ Pivot Table/Edit Pivot Table/NET Standard/Edit Pivot Table/Edit Pivot Table.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34310.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Edit Pivot Table", "Edit Pivot Table\Edit Pivot Table.csproj", "{6A87150C-3AB9-4F73-943D-880A5091DBCD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6A87150C-3AB9-4F73-943D-880A5091DBCD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6A87150C-3AB9-4F73-943D-880A5091DBCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6A87150C-3AB9-4F73-943D-880A5091DBCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6A87150C-3AB9-4F73-943D-880A5091DBCD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {5438F081-8D9A-4D05-8234-C220E1FE4339}
+	EndGlobalSection
+EndGlobal
