+++ Pivot Table/Edit Pivot Table/NET Standard/Edit Pivot Table/Edit Pivot Table/Program.cs
@@ -1,0 +1,47 @@
+﻿using System.IO;
+using Syncfusion.XlsIO;
+using Syncfusion.XlsIO.Implementation.PivotTables;
+
+namespace Edit_Pivot_Table
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            using (ExcelEngine excelEngine = new ExcelEngine())
+            {
+                IApplication application = excelEngine.Excel;
+                application.DefaultVersion = ExcelVersion.Xlsx;
+                FileStream inputStream = new FileStream("../../../Data/InputTemplate.xlsx", FileMode.Open, FileAccess.Read);
+                IWorkbook workbook = application.Workbooks.Open(inputStream);
+                IWorksheet worksheet = workbook.Worksheets[1];
+
+                //Accessing the pivot table in the worksheet
+                IPivotTable pivotTable = worksheet.PivotTables[0];
+
+                //Layout the pivot table to set the values to the worksheet
+                pivotTable.Layout();
+
+                //Set Text in cell B2
+                worksheet.Range["B2"].Text = "William";
+
+                #region Save
+                //Saving the workbook
+                FileStream outputStream = new FileStream("Output.xlsx", FileMode.Create, FileAccess.Write);
+                workbook.SaveAs(outputStream);
+                #endregion
+
+                //Dispose streams
+                outputStream.Dispose();
+                inputStream.Dispose();
+
+                System.Diagnostics.Process process = new System.Diagnostics.Process();
+                process.StartInfo = new System.Diagnostics.ProcessStartInfo("Output.xlsx")
+                {
+                    UseShellExecute = true
+                };
+                process.Start();
+            }
+        }
+    }
+}
