+++ Pivot Table/Outline Layout/NET Standard/Outline Layout/Outline Layout.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34310.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Outline Layout", "Outline Layout\Outline Layout.csproj", "{7DC5953B-A556-4CFA-B4AE-A02194BF06D1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7DC5953B-A556-4CFA-B4AE-A02194BF06D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7DC5953B-A556-4CFA-B4AE-A02194BF06D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7DC5953B-A556-4CFA-B4AE-A02194BF06D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7DC5953B-A556-4CFA-B4AE-A02194BF06D1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {89A0827A-6E32-4814-80A3-786F7D630F7A}
+	EndGlobalSection
+EndGlobal
