--- Pivot Table/Pivot Layout/NET Standard/Pivot Layout/Pivot Layout.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31729.503
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Pivot Layout", "Pivot Layout\Pivot Layout.csproj", "{B3538511-0402-4EA1-9A71-98DD48DD2D58}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B3538511-0402-4EA1-9A71-98DD48DD2D58}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B3538511-0402-4EA1-9A71-98DD48DD2D58}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B3538511-0402-4EA1-9A71-98DD48DD2D58}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B3538511-0402-4EA1-9A71-98DD48DD2D58}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {5E852D46-93F0-476B-AC81-E72656F5156C}
-	EndGlobalSection
-EndGlobal
