--- Pivot Table/Pivot Layout/NET Standard/Pivot Layout/Pivot Layout/Pivot Layout.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <RootNamespace>Pivot_Layout</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Syncfusion.XlsIO.NET" Version="19.4.0.38" />
-  </ItemGroup>
-
-</Project>
