+++ Pivot Table/Remove Pivot Table/NET Standard/Remove Pivot Table/Remove Pivot Table.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34310.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Remove Pivot Table", "Remove Pivot Table\Remove Pivot Table.csproj", "{FB86801B-ABFC-470B-B1FD-B6016DD39BF7}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{FB86801B-ABFC-470B-B1FD-B6016DD39BF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB86801B-ABFC-470B-B1FD-B6016DD39BF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB86801B-ABFC-470B-B1FD-B6016DD39BF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB86801B-ABFC-470B-B1FD-B6016DD39BF7}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {6CB1508A-307C-430C-A437-745C8BEC48F3}
+	EndGlobalSection
+EndGlobal
