+++ Pivot Table/Tabular Layout/NET Standard/Tabular Layout/Tabular Layout.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.9.34310.174
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tabular Layout", "Tabular Layout\Tabular Layout.csproj", "{9F6F6C3D-49B0-4BFF-9DE1-5E4C8B935534}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9F6F6C3D-49B0-4BFF-9DE1-5E4C8B935534}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9F6F6C3D-49B0-4BFF-9DE1-5E4C8B935534}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9F6F6C3D-49B0-4BFF-9DE1-5E4C8B935534}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9F6F6C3D-49B0-4BFF-9DE1-5E4C8B935534}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {C1A9BCFF-BED3-4791-846B-9150A98B2A26}
+	EndGlobalSection
+EndGlobal
