+++ Pivot Table/Tabular Layout/NET Standard/Tabular Layout/Tabular Layout/Program.cs
@@ -1,0 +1,44 @@
+﻿using Syncfusion.XlsIO;
+using System.IO;
+
+namespace Tabular_Layout
+{
+    class Program
+    {
+        public static void Main(string[] args)
+        {
+            using (ExcelEngine excelEngine = new ExcelEngine())
+            {
+                IApplication application = excelEngine.Excel;
+                application.DefaultVersion = ExcelVersion.Xlsx;
+                FileStream inputStream = new FileStream("../../../Data/InputTemplate.xlsx", FileMode.Open, FileAccess.Read);
+                IWorkbook workbook = application.Workbooks.Open(inputStream);
+                IWorksheet worksheet = workbook.Worksheets[1];
+                IPivotTable pivotTable = worksheet.PivotTables[0];
+
+                //Set PivotTableRowLayout
+                pivotTable.Options.RowLayout = PivotTableRowLayout.Tabular;
+
+                //Set BuiltInStyle
+                pivotTable.BuiltInStyle = PivotBuiltInStyles.PivotStyleMedium9;
+
+                #region Save
+                //Saving the workbook
+                FileStream outputStream = new FileStream("Output.xlsx", FileMode.Create, FileAccess.Write);
+                workbook.SaveAs(outputStream);
+                #endregion
+
+                //Dispose streams
+                outputStream.Dispose();
+                inputStream.Dispose();
+
+                System.Diagnostics.Process process = new System.Diagnostics.Process();
+                process.StartInfo = new System.Diagnostics.ProcessStartInfo("Output.xlsx")
+                {
+                    UseShellExecute = true
+                };
+                process.Start();
+            }
+        }
+    }
+}
